Update implementation of moment calculations
The SKEstimator constructor  now uses Float64 for M and N, but requires
that M, N, and M*N*d are integers.  M and N must be integers because
their fields in SKEstimator are Int.  M*N*d must be integer because the
gamma function is implemented as an integer factorial.

A convenience constructor allows Integers to be passed for M and N.