Return error along with Pearson probability curves
When calling `pearson_type_iii` or `pearson_type_vi` to get a Pearson
Type III or Pearson Type VI approximation of the distribution of an
`SKEstimator', the relative error in the fourth central moment is now
also returned.  The relative error in the fourth central moment can be
used to determine which Pearson approximation will be best suited for
determining outlier thresholds for a given `SKEstimator`.  A lower error
in the fourth central moment implies a better fit to the tails of the
distribution which implies a more accurate outlier threshold.

This change necessitates the removal of the trivial 3-arg convenience
methods `pearson_type_iii(M,N,d)` and `pearson_type_vi(M,n,d)`.