Add cdf and quantile support for PearsonTypeIV
The CDF of a PearsonTypeIV is implemented as an approximated function
that is unique to the specific PearsonTypeIV instance (and domain).
The `quantile(d,x)` method solves for `x` such that `cdf(d, x) = q`.
With the addition of `quantile` support, the `thresholds` function can
now return thresholds for PearsonTypeIV instances.